[REFACTOR] #32 - Model 이름 변경, Package 구조 변경
- LoginResponse -> TokenModel로 변경
- UserInfo -> UserModel로 변경
- UserStarred -> StarredModel로 변경
- Repositories 패키지를 data 패키지 아래로 이동
- model 패키지 이름 models로 변경
- viewmodels 패키지를 ui 패키지 아래로 이동
