+++ app/src/main/java/com/example/android_repo_05/data/models/UserProfileResponse.kt
@@ -1,4 +1,4 @@
-package com.example.android_repo_05.data.model
+package com.example.android_repo_05.data.models
 
 import com.google.gson.annotations.SerializedName
