+++ app/src/main/java/com/example/android_repo_05/paging/IssuePagingSource.kt
@@ -2,10 +2,8 @@ package com.example.android_repo_05.paging
 
 import androidx.paging.PagingSource
 import androidx.paging.PagingState
-import com.example.android_repo_05.data.model.IssueResponse
+import com.example.android_repo_05.data.models.IssueResponse
 import com.example.android_repo_05.retrofit.GithubApiInstance.retrofit
-import retrofit2.await
-import java.lang.Exception
 
 private const val STARTING_PAGE_INDEX = 1
 private const val NETWORK_PAGE_SIZE = 10
