+++ app/src/main/java/com/example/android_repo_05/retrofit/GithubApi.kt
@@ -1,11 +1,11 @@
 package com.example.android_repo_05.retrofit
 
 import com.example.android_repo_05.BuildConfig
-import com.example.android_repo_05.data.model.IssueResponse
-import com.example.android_repo_05.data.model.LoginResponse
-import com.example.android_repo_05.data.model.UserInfo
-import com.example.android_repo_05.data.model.UserProfileResponse
-import com.example.android_repo_05.data.model.UserStarred
+import com.example.android_repo_05.data.models.IssueResponse
+import com.example.android_repo_05.data.models.TokenModel
+import com.example.android_repo_05.data.models.UserModel
+import com.example.android_repo_05.data.models.UserProfileResponse
+import com.example.android_repo_05.data.models.StarredModel
 import com.example.android_repo_05.others.Constants
 import retrofit2.Response
 import retrofit2.http.*
@@ -18,10 +18,10 @@ interface GithubApi {
         @Field("client_id") clientId: String = BuildConfig.GITHUB_CLIENT_ID,
         @Field("client_secret") clientSecrets: String = BuildConfig.GITHUB_CLIENT_SECRETS,
         @Field("code") accessCode: String
-    ): Response<LoginResponse>
+    ): Response<TokenModel>
 
     @GET("user")
-    suspend fun getUserInfo(): Response<UserInfo>
+    suspend fun getUserInfo(): Response<UserModel>
 
     @GET("user/issues")
     suspend fun getIssues(
@@ -31,7 +31,7 @@ interface GithubApi {
     ): ArrayList<IssueResponse>
 
     @GET("user/starred")
-    suspend fun getUserStarred(): Response<List<UserStarred>>
+    suspend fun getUserStarred(): Response<List<StarredModel>>
 
     @GET("user")
     suspend fun getProfileUrl() : Response<UserProfileResponse>
