+++ app/src/main/java/com/example/android_repo_05/ui/activities/LoginActivity.kt
@@ -6,13 +6,13 @@ import android.view.View
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.ViewModelProvider
 import com.example.android_repo_05.R
-import com.example.android_repo_05.data.model.LoginResponse
-import com.example.android_repo_05.data.model.ResponseState
+import com.example.android_repo_05.data.models.TokenModel
+import com.example.android_repo_05.data.models.ResponseState
 import com.example.android_repo_05.databinding.ActivityLoginBinding
 import com.example.android_repo_05.others.Utils
-import com.example.android_repo_05.repositories.TokenRepository
-import com.example.android_repo_05.viewmodels.LoginViewModel
-import com.example.android_repo_05.viewmodels.AppViewModelFactory
+import com.example.android_repo_05.data.repositories.TokenRepository
+import com.example.android_repo_05.ui.viewmodels.LoginViewModel
+import com.example.android_repo_05.ui.viewmodels.AppViewModelFactory
 import com.google.android.material.snackbar.Snackbar
 
 class LoginActivity : AppCompatActivity() {
@@ -59,7 +59,7 @@ class LoginActivity : AppCompatActivity() {
     }
 
     private fun setObserver() {
-        loginViewModel.loginResponse.observe(this) { responseState ->
+        loginViewModel.tokenModel.observe(this) { responseState ->
             when (responseState) {
                 is ResponseState.Success -> handleLoginSuccess(responseState)
                 is ResponseState.Error -> handleLoginFailure()
@@ -75,7 +75,7 @@ class LoginActivity : AppCompatActivity() {
         즉, response.tokenType이 blanck가 아니면 remote에서 access token을 받아 온 것이기 때문에
         data store에 저장해야 함
     */
-    private fun handleLoginSuccess(state: ResponseState<LoginResponse>) {
+    private fun handleLoginSuccess(state: ResponseState<TokenModel>) {
         binding.cpiLogin.visibility = View.INVISIBLE
         state.data?.let { response ->
             if (response.tokenType.isNotBlank()) {
