+++ app/src/main/java/com/example/android_repo_05/ui/viewmodels/LoginViewModel.kt
@@ -1,27 +1,27 @@
-package com.example.android_repo_05.viewmodels
+package com.example.android_repo_05.ui.viewmodels
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.android_repo_05.base.CustomApplication.Companion.setAccessToken
-import com.example.android_repo_05.data.model.LoginResponse
-import com.example.android_repo_05.data.model.ResponseState
-import com.example.android_repo_05.repositories.TokenRepository
+import com.example.android_repo_05.data.models.TokenModel
+import com.example.android_repo_05.data.models.ResponseState
+import com.example.android_repo_05.data.repositories.TokenRepository
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import retrofit2.Response
 
 class LoginViewModel(private val repository: TokenRepository) : ViewModel() {
-    private var _loginResponse: MutableLiveData<ResponseState<LoginResponse>> = MutableLiveData()
-    val loginResponse: LiveData<ResponseState<LoginResponse>> get() = _loginResponse
+    private var _tokenModel: MutableLiveData<ResponseState<TokenModel>> = MutableLiveData()
+    val tokenModel: LiveData<ResponseState<TokenModel>> get() = _tokenModel
 
     fun getAccessTokenFromRemote(code: String) = viewModelScope.launch(Dispatchers.IO) {
-        _loginResponse.postValue(ResponseState.Loading())
-        _loginResponse.postValue(handleLoginResponse(repository.getAccessTokenFromRemote(code)))
+        _tokenModel.postValue(ResponseState.Loading())
+        _tokenModel.postValue(handleLoginResponse(repository.getAccessTokenFromRemote(code)))
     }
 
-    private fun handleLoginResponse(response: Response<LoginResponse>): ResponseState<LoginResponse> {
+    private fun handleLoginResponse(response: Response<TokenModel>): ResponseState<TokenModel> {
         if (response.isSuccessful) {
             response.body()?.let { result ->
                 return ResponseState.Success(result)
