+++ app/src/main/java/com/example/android_repo_05/ui/viewmodels/UserInfoViewModel.kt
@@ -1,26 +1,26 @@
-package com.example.android_repo_05.viewmodels
+package com.example.android_repo_05.ui.viewmodels
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.android_repo_05.data.model.ResponseState
-import com.example.android_repo_05.data.model.UserInfo
-import com.example.android_repo_05.data.model.UserStarred
-import com.example.android_repo_05.repositories.TokenRepository
+import com.example.android_repo_05.data.models.ResponseState
+import com.example.android_repo_05.data.models.UserModel
+import com.example.android_repo_05.data.models.StarredModel
+import com.example.android_repo_05.data.repositories.TokenRepository
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import retrofit2.Response
 
 class UserInfoViewModel(private val repository: TokenRepository) : ViewModel() {
-    private var _userInfo: MutableLiveData<ResponseState<UserInfo>> = MutableLiveData()
-    val userInfo: LiveData<ResponseState<UserInfo>> get() = _userInfo
+    private var _userModel: MutableLiveData<ResponseState<UserModel>> = MutableLiveData()
+    val userModel: LiveData<ResponseState<UserModel>> get() = _userModel
     private var _starredCount: MutableLiveData<ResponseState<Int>> = MutableLiveData()
     val starredCount: LiveData<ResponseState<Int>> get() = _starredCount
 
     fun getUserInfoFromRemote() = viewModelScope.launch(Dispatchers.IO) {
-        _userInfo.postValue(ResponseState.Loading())
-        _userInfo.postValue(handleUserInfoResponse(repository.getUserInfoFromRemote()))
+        _userModel.postValue(ResponseState.Loading())
+        _userModel.postValue(handleUserInfoResponse(repository.getUserInfoFromRemote()))
     }
 
     fun getUserStarredFromRemote() = viewModelScope.launch(Dispatchers.IO) {
@@ -28,7 +28,7 @@ class UserInfoViewModel(private val repository: TokenRepository) : ViewModel() {
         _starredCount.postValue(handleUserStarredResponse(repository.getUserStarred()))
     }
 
-    private fun handleUserInfoResponse(response: Response<UserInfo>): ResponseState<UserInfo> {
+    private fun handleUserInfoResponse(response: Response<UserModel>): ResponseState<UserModel> {
         if (response.isSuccessful) {
             response.body()?.let { result ->
                 return ResponseState.Success(result)
@@ -37,7 +37,7 @@ class UserInfoViewModel(private val repository: TokenRepository) : ViewModel() {
         return ResponseState.Error(response.message())
     }
 
-    private fun handleUserStarredResponse(response: Response<List<UserStarred>>): ResponseState<Int> {
+    private fun handleUserStarredResponse(response: Response<List<StarredModel>>): ResponseState<Int> {
         if (response.isSuccessful) {
             response.body()?.let { result ->
                 return ResponseState.Success(result.size)
